initial 1.3.2 commit and 1.3.1

+++ Utils.cs
@@ -146,9 +146,9 @@ namespace Retailier
 					{
 						value = element.GetInt32();
 					}
-					else if (element.GetType() == typeof(double))
+					else if (element.GetType() == typeof(float))
 					{
-						value = ((float)element.GetDouble());
+						value = (float) element.GetDouble();
 					}
 					break;
